Koala on the wheel
Completely rewritten storage structures
Fixed a bug in the Rung Koute, too high order of the system
The code is translated into koaly calculations on the wheel
The flight in the air is considered by a very simplified model. The wheel moves independently of the body. Body is displaced to the same vector